refactor: JanggiGame 생성 책임을 GameInitializer로 분리
